moved s3Storage to xpub-client
